@@ -12,16 +12,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.6",
-    "@angular/common": "^17.3.6",
-    "@angular/compiler": "^17.3.6",
-    "@angular/core": "^17.3.6",
-    "@angular/forms": "^17.3.6",
-    "@angular/platform-browser": "^17.3.6",
-    "@angular/platform-browser-dynamic": "^17.3.6",
-    "@angular/router": "^17.3.6",
-    "@angular/youtube-player": "^17.3.6",
-    "@asymmetrik/ngx-leaflet": "^17.0.0",
+    "@angular/animations": "^18.0.2",
+    "@angular/common": "^18.0.2",
+    "@angular/compiler": "^18.0.2",
+    "@angular/core": "^18.0.2",
+    "@angular/forms": "^18.0.2",
+    "@angular/platform-browser": "^18.0.2",
+    "@angular/platform-browser-dynamic": "^18.0.2",
+    "@angular/router": "^18.0.2",
+    "@angular/youtube-player": "^18.0.2",
+    "@asymmetrik/ngx-leaflet": "^18.0.1",
     "leaflet": "^1.9.4",
     "primeicons": "^7.0.0",
     "primeng": "^17.13.0",
@@ -30,9 +30,9 @@
     "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.6",
-    "@angular/cli": "~17.3.6",
-    "@angular/compiler-cli": "^17.3.6",
+    "@angular-devkit/build-angular": "^18.0.3",
+    "@angular/cli": "~18.0.3",
+    "@angular/compiler-cli": "^18.0.2",
     "@types/eslint": "^8.56.9",
     "@types/jasmine": "~5.1.4",
     "@types/leaflet": "^1.9.9",
