@@ -4,7 +4,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { CardModule } from 'primeng/card';
 import { MessageService } from 'primeng/api';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
@@ -13,15 +13,8 @@ import { TooltipModule } from 'primeng/tooltip';
 
 @NgModule({
   declarations: [AppComponent, HomeComponent],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    HttpClientModule,
-    CardModule,
-    BrowserAnimationsModule,
-    TooltipModule,
-  ],
-  providers: [MessageService],
   bootstrap: [AppComponent],
+  imports: [BrowserModule, AppRoutingModule, CardModule, BrowserAnimationsModule, TooltipModule],
+  providers: [MessageService, provideHttpClient(withInterceptorsFromDi())],
 })
 export class AppModule {}
